// hub/frontend/lib/api/generated/model/organizationRead.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.93
 */

export interface OrganizationRead {
  createdAt: string;
  currentUserRole: string;
  iconUrl: string;
  id: string;
  name: string;
  slug: string;
  status: string;
  updatedAt: string;
}