// hub/frontend/lib/api/generated/model/submissionRead.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.84
 */
import type { SubmissionReadServerJson } from './submissionReadServerJson';
import type { SubmissionReadStatus } from './submissionReadStatus';
import type { SubmissionReadSubmissionType } from './submissionReadSubmissionType';
import type { SubmissionReadValidationResult } from './submissionReadValidationResult';

export interface SubmissionRead {
  approvedAt?: string | null;
  approverUserId?: string | null;
  createdAt: string;
  id: string;
  name: string;
  ownerOrganizationId?: string | null;
  ownerUserId?: string | null;
  publishedServerVersionId?: string | null;
  rejectionMessage: string;
  serverJson: SubmissionReadServerJson;
  status: SubmissionReadStatus;
  submissionType: SubmissionReadSubmissionType;
  submittedAt?: string | null;
  submitterUserId: string;
  updatedAt: string;
  validationResult: SubmissionReadValidationResult;
  version: string;
}